TASK-60 [Feat]: Create File Header.tsx & Layout.tsx
