+++ .eslintrc.json
@@ -21,16 +21,6 @@
   },
   "plugins": ["react", "@typescript-eslint", "react-refresh", "prettier"],
   "rules": {
-    //"sort-imports": [
-    //  "error",
-    //  {
-    //    "ignoreCase": false,
-    //    "ignoreDeclarationSort": false,
-    //    "ignoreMemberSort": false,
-    //    "memberSyntaxSortOrder": ["none", "all", "multiple", "single"],
-    //    "allowSeparatedGroups": false
-    //  }
-    //],
     "react/function-component-definition": [
       2,
       {
