+++ src/components/@shared/Layout/Layout.tsx
@@ -1,0 +1,11 @@
+import React, { PropsWithChildren } from 'react';
+import Header from '../Common/Header';
+
+export default function Layout({ children }: PropsWithChildren) {
+  return (
+    <>
+      <Header />
+      {children}
+    </>
+  );
+}
