--- src/core/types/RootContext.d.ts
@@ -1,2 +1,0 @@
-export type UserType = UserServiceReponseDto;
-export type LoginType = (body: LoginRequestDto) => Promise<void>;
